add new notifications to ide (#4406)
