+++ packages/app/src/app/components/Overlay/Overlay.tsx
@@ -18,7 +18,6 @@ interface IOverlayProps {
   children: (handleOpen: () => void) => React.ReactNode;
   content: React.ComponentType;
   width?: number;
-  propagate?: boolean;
 }
 
 const POPOVER_WIDTH = 390;
@@ -32,7 +31,6 @@ export const Overlay: React.FC<IOverlayProps> = ({
   children,
   content: Content,
   width = POPOVER_WIDTH,
-  propagate = false,
 }) => {
   const [open, setOpen] = useState(isOpen === undefined ? false : isOpen);
   const isControlled = isOpen !== undefined;
@@ -115,11 +113,10 @@ export const Overlay: React.FC<IOverlayProps> = ({
     }
   };
 
+  const multiplier = width < POPOVER_WIDTH ? 0.9 : 0.75;
+
   return (
-    <Container
-      onMouseDown={e => (propagate ? noop : e.stopPropagation())}
-      ref={element}
-    >
+    <Container ref={element}>
       {children(handleOpen)}
 
       {openState && (
@@ -132,11 +129,15 @@ export const Overlay: React.FC<IOverlayProps> = ({
                 position: 'absolute',
                 zIndex: 10,
                 overflow: 'hidden',
-                boxShadow: '0 3px 3px rgba(0, 0, 0, 0.3)',
+                boxShadow:
+                  '0px 4px 4px rgba(0,0,0,0.12), 0px 16px 32px rgba(0,0,0,0.24)',
                 height: 'auto',
                 borderRadius: 4,
                 top: bounds.current.top + bounds.current.height + 8,
-                left: bounds.current.left + bounds.current.width - width * 0.75,
+                left:
+                  bounds.current.left +
+                  bounds.current.width -
+                  width * multiplier,
                 width,
               }}
               transition={{
