+++ packages/app/src/app/pages/Sandbox/Editor/Header/Actions.tsx
@@ -4,6 +4,7 @@ import css from '@styled-system/css';
 import { useOvermind } from 'app/overmind';
 import { UserMenu } from 'app/pages/common/UserMenu';
 import React, { useEffect, useState } from 'react';
+import { Notifications } from 'app/components/Notifications';
 
 import {
   EmbedIcon,
@@ -182,6 +183,7 @@ export const Actions = () => {
       >
         Create Sandbox
       </Button>
+      {hasLogIn && <Notifications />}
       {hasLogIn ? (
         <UserMenu>
           {user?.experiments.collaborator ? (
